+++ frontend/src/user/components/auth/index.tsx
@@ -1,22 +1,8 @@
-import { FC, useState } from "react";
-import { useNavigate } from "react-router";
-import { useLogoutMutation } from "../../../store/api/authApi";
+import { FC } from "react";
+import { useLogout } from "../../../common/hooks/useLogout";
 
 export const AuthPage: FC = () => {
-  const navigate = useNavigate();
-  const goToLoginPage = () => navigate("/login");
-  const [error, setError] = useState<string | null>(null);
-
-  const [mutation] = useLogoutMutation();
-
-  const handleLogout = async () => {
-    try {
-      await mutation().unwrap();
-      goToLoginPage();
-    } catch (error: any) {
-      setError(error?.data?.error || "予期せぬエラーが発生しました");
-    }
-  };
+  const { handleLogout, error } = useLogout();
 
   return (
     <>
